Pro-Forma Fix Delete Confirmation

+++ src/app/pages/apps/pro-forma-entries/pro-forma-entries-list/pro-forma-entries-list.component.ts
@@ -18,6 +18,7 @@ import { PortalService } from 'src/app/core/services/portal.service';
 import { IProFormaEntry } from 'src/app/core/models/pro-forma-entry';
 import { ProFormaEntryStatusEnum } from 'src/app/core/enums/pro-forma-entry-status.enum';
 import { Router } from '@angular/router';
+import { DialogComponent } from '../../dialog/dialog.component';
 
 @UntilDestroy()
 @Component({
@@ -164,13 +165,33 @@ export class ProFormaEntriesListComponent implements OnInit, OnDestroy, AfterVie
   }
 
   delete(proFormaEntry: any): void {
-    this._portalService.deleteProFormaEntry(proFormaEntry.id)
-    .pipe(takeUntil(this._onDestroy$))
-    .subscribe(libraryData => {
-      if (!libraryData) {
-        return;
+
+
+    const dialogRef = this._dialog.open(DialogComponent, {
+      data: {
+        cancelButtonLabel: "Cancel",
+        confirmButtonLabel: "Continue delete pro-forma entry",
+        dialogHeader: "Delete pro-forma entry",
+        dialogContent: "Are you sure you would like to delete this pro-forma entry?",
+        callbackMethod: function() {
+         
+        }
+      },
+      width: '700px'
+    });
+
+    
+    dialogRef.afterClosed().subscribe(result => {
+      if(result != undefined) {
+        this._portalService.deleteProFormaEntry(proFormaEntry.id)
+        .pipe(takeUntil(this._onDestroy$))
+        .subscribe(libraryData => {
+          if (!libraryData) {
+            return;
+          }
+        
+        });
       }
-     
     });
   }
 }
